+++ CHANGES.md
@@ -1,3 +1,9 @@
+# 0.0.4
+
+- Update the tools.deps dependency to 0.5.452
+- The badigeon.compile/compile function now takes an optional :classpath argument
+- Add the possibility to compute a classpath string using a tools.deps dependency map and optional aliases. 
+
 # 0.0.3
 
 - Add the possibility to execute shell commands in a separate process.